resolve comments from jpdeleon & enzoampil
